requirements.txt
requests
python-dotenv
zeep